Share, Together Update 수정

+++ src/main/java/org/devkirby/hanimman/controller/ShareController.java
@@ -34,7 +34,10 @@ public class ShareController {
         }else if(shareDTO.getContent().length() > 65535){
             throw new IllegalArgumentException("내용의 길이는 65535자 이하여야 합니다. 현재 길이: "
                     + shareDTO.getContent().length());
-        }else{
+        }else if(shareDTO.getFiles().size()>10){
+            throw new IllegalArgumentException("이미지는 최대 10개까지 업로드할 수 있습니다. 현재 이미지 개수: "
+                    + shareDTO.getFiles().size());
+        } else {
             shareDTO.setUserId(loginUser.getId());
             shareService.create(shareDTO);
             map.put("code", 200);
@@ -49,14 +52,36 @@ public class ShareController {
     }
 
     @PutMapping("/{id}")
-    public void updateShare(@PathVariable Integer id, @RequestBody ShareDTO shareDTO) {
+    public Map<String, Object> updateShare(@PathVariable Integer id, @RequestBody ShareDTO shareDTO, @AuthenticationPrincipal User loginUser) throws IOException {
+        Map<String, Object> map = new HashMap<>();
         shareDTO.setId(id);
-        shareService.update(shareDTO);
+        if(!loginUser.getId().equals(shareDTO.getUserId())){
+            throw new IllegalArgumentException("본인이 작성한 게시글만 수정할 수 있습니다.");
+        }if(shareDTO.getTitle().length() > 255 || shareDTO.getTitle().isEmpty()){
+            throw new IllegalArgumentException("제목의 길이는 1자 이상, 255자 이하여야 합니다. 현재 길이: "
+                    + shareDTO.getTitle().length());
+        }else if(shareDTO.getContent().length() > 65535){
+            throw new IllegalArgumentException("내용의 길이는 65535자 이하여야 합니다. 현재 길이: "
+                    + shareDTO.getContent().length());
+        }else{
+            shareService.update(shareDTO);
+            map.put("code", 200);
+            map.put("msg", "나눠요 게시글 수정에 성공했습니다.");
+            return map;
+        }
     }
 
     @DeleteMapping("/{id}")
-    public void deleteShare(@PathVariable Integer id) {
-        shareService.delete(id);
+    public Map<String, Object> deleteShare(@PathVariable Integer id, @AuthenticationPrincipal User loginUser) {
+        Map<String, Object> map = new HashMap<>();
+        if(!loginUser.getId().equals(shareService.read(id, loginUser).getUserId())){
+            throw new IllegalArgumentException("본인이 작성한 게시글만 삭제할 수 있습니다.");
+        }else{
+            shareService.delete(id);
+            map.put("code", 200);
+            map.put("msg", "나눠요 게시글 삭제에 성공했습니다.");
+            return map;
+        }
     }
 
     @GetMapping
