+++ src/main/java/org/devkirby/hanimman/controller/TogetherController.java
@@ -34,7 +34,10 @@ public class TogetherController {
         }else if(togetherDTO.getContent().length() > 1000){
             throw new IllegalStateException("내용의 길이는 65535자 이하여야 합니다. 현재 길이 : " +
                     + togetherDTO.getContent().length());
-        }else{
+        }else if(togetherDTO.getFiles().size()>10){
+            throw new IllegalStateException("이미지는 최대 10개까지 업로드할 수 있습니다. 현재 이미지 개수 : " +
+                    + togetherDTO.getFiles().size());
+        } else {
             togetherDTO.setUserId(loginUser.getId());
             togetherService.create(togetherDTO);
             map.put("code", 200);
@@ -49,14 +52,36 @@ public class TogetherController {
     }
 
     @PutMapping("/{id}")
-    public void updateTogether(@PathVariable Integer id, @RequestBody TogetherDTO togetherDTO) {
+    public Map<String, Object> updateTogether(@PathVariable Integer id, @RequestBody TogetherDTO togetherDTO, @AuthenticationPrincipal User loginUser) throws IOException {
+        Map<String, Object> map = new HashMap<>();
         togetherDTO.setId(id);
-        togetherService.update(togetherDTO);
+        if(!loginUser.getId().equals(togetherDTO.getUserId())) {
+            throw new IllegalArgumentException("본인이 작성한 게시글만 수정할 수 있습니다.");
+        }else if(togetherDTO.getTitle().length() > 255 || togetherDTO.getTitle().isEmpty()){
+            throw new IllegalStateException("제목의 길이는 1자 이상, 255자 이하여야 합니다. 현재 길이 : " +
+                    + togetherDTO.getTitle().length());
+        }else if(togetherDTO.getContent().length() > 65535){
+            throw new IllegalStateException("내용의 길이는 65535자 이하여야 합니다. 현재 길이 : " +
+                    + togetherDTO.getContent().length());
+        }else {
+            togetherService.update(togetherDTO);
+            map.put("code", 200);
+            map.put("msg", "같이가요 게시글 수정에 성공했습니다.");
+            return map;
+        }
     }
 
     @DeleteMapping("/{id}")
-    public void deleteTogether(@PathVariable Integer id) {
-        togetherService.delete(id);
+    public Map<String, Object> deleteTogether(@PathVariable Integer id, @AuthenticationPrincipal User loginUser) {
+        Map<String, Object> map = new HashMap<>();
+        if(!loginUser.getId().equals(togetherService.read(id, loginUser).getUserId())) {
+            throw new IllegalArgumentException("본인이 작성한 게시글만 삭제할 수 있습니다.");
+        }else{
+            togetherService.delete(id);
+            map.put("code", 200);
+            map.put("msg", "같이가요 게시글 삭제에 성공했습니다.");
+            return map;
+        }
     }
 
     @GetMapping
